Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
httpx==0.27.2
livekit-agents==1.2.12
livekit-plugins-openai==1.2.12
livekit-plugins-deepgram==1.2.12
livekit-plugins-cartesia==1.2.12
livekit-plugins-silero==1.2.12
openai>=1.99.2
gunicorn==21.2.0
